Add a comment to sushy_driver SSH pub key
Adding this comment, especially when injecting the pub key in the
hypervisor authorized_keys, allows to then filter out that specific key
when we generate the cloud-init "user-data".

This is needed in order to ensure we can re-run the deployment, for
instance when doing a scale-up or scale-down, or just re-running for
whatever reason:
since the sushy_driver key is injected "late" to the hypervisor
authorized_keys via reproducer role, it would then modify the user-data
generated during the libvirt_manager stage, leading to a file change,
leading eventually to a failure since it would re-generate the connfig
drive ISO (and this is prevented).

This patch ensures indempotency of the runs, by ensuring the
"late" sushy_driver ssh key isn't injected.

We also take the opportunity to remove the sushy_driver pub key via the
reproducer-clean.yml, ensuring we don't get an ever-growing file on the
hypervisor.