Change how tempest is configured in test_operator role
Until now, whenever someone wanted to add a new option to their tempest
configuration, they had to add it directly in the `defaults.yml`.

This change proposes another approach, like we have for tobiko
configuration: a default set of configuration entries, and a parameter
allowing to override it.

In order to avoid breaking existing configurations, the default value is
still consuming the various parameters introduced over time - meaning
most of the existing jobs shouldn't be affected.