Rework inventory generation
First we move the inventory files generation to a far earlier stage.
While it doesn't change anything on its content, having the inventory
generated earlier helps debugging it without the need to create any
actual workload (hacky approach I used to validate the generated
inventories).

Then, the data used to generate the inventories is updated, in order to
ensure we're still puting the nodes under the right group.
This is especially important for the OCP cluster, when we have workers
in the picture.

Before this change, the workers were nested under the "ocps" group,
leaving the "ocp_workers" empty.

This change corrects the situation, leading the generated inventories to
look like this:
```yaml
ocps:
  hosts:
    ocp-master-0:
      ansible_host: master-0.utility
      ansible_user: core
      ansible_ssh_common_args: '-o StrictHostKeyChecking=no'
      ansible_ssh_private_key_file: ~/.ssh/devscripts_key
ocp_workers:
  hosts:
    ocp-worker-0:
      ansible_host: worker-0.utility
      ansible_user: core
      ansible_ssh_common_args: '-o StrictHostKeyChecking=no'
      ansible_ssh_private_key_file: ~/.ssh/devscripts_key
computes:
  hosts:
    compute-xio2mlri-0:
      ansible_host: compute-xio2mlri-0.utility
      ansible_user: zuul
      ansible_ssh_common_args: '-o StrictHostKeyChecking=no'
      ansible_ssh_private_key_file: ~/.ssh/id_cifw
controllers:
  hosts:
    controller-0:
      ansible_host: controller-0.utility
      ansible_user: zuul
      ansible_ssh_common_args: '-o StrictHostKeyChecking=no'
      ansible_ssh_private_key_file: ~/.ssh/id_cifw
```