Run ceph update if part of the deployment
In case we perform a minor update job where Ceph is deployed, as part of
the minor update procedure we might want to update Ceph (according to a
given container tag) to verify the combination between the target Ceph
version and the target RHOSO version is still healthy.
This patch introduces the tasks that are supposed to call the Ceph
update before openstack:
- it results in a noop in case the ceph container image tag is the <= of
  the current deployed Ceph cluster
- it is skipped by default if no "ceph_update" variable is passed or it
  is set to false
- it can be skipped by passing --skip-tags ceph_update to the playbook

By doing this we have the ability to trigger a Ceph update and make sure
we continue to test the combination between the two projects, but at the
same time we are not introducing an hard dependency as it can easily be
skipped.

Jira: https://issues.redhat.com/browse/OSPRH-9697

Signed-off-by: Francesco Pantano <fpantano@redhat.com>