Replace 'operator' filter with cifmw_set_openstack_containers_excluded_envs
Replace condition ('operator' not in _container) with more understandable
and predictable list of ENV vars that should not be modifed by the role.
Previous filter was created in assumption that container names follow
some pattern, but as discovered recently they are diffreneces in upstream
versus downstream naming. ENVs names are the same, because they are defined
in operators and we should use them instead.