Make sure set_openstack_container is less depedent on instal_yamls
Cleanup set_openstack_container role by setting default values and omitting
setting fact if install_yamls was not used during deployment.

This is first atempt of reducing dependency of set_openstack_container role
from install_yamls. The work that is left for later is for updating
envs in operators other than openstack-operator. Task that waits
for an operator update to complete will only work in install_yaml deployed
enviroments and is skipped in architecture_scenario context.