Use Ceph monitoring (not RGW) network for Dashboard
Use the existing `cifmw_cephadm_monitoring_network`
parameter to configure the Ceph dashboard instead of
the `cifmw_cephadm_rgw_network` parameter. This was
not detected earlier because the Ceph playbook set them
both to the Ceph public_network, but we recently added
a feature to run RGW on a separate network [1].
Also, add retries/delay to the dashboard HTTP check
and mange if HTTPS should be used.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2301

Jira: https://issues.redhat.com/browse/OSPCIX-506

Signed-off-by: John Fulton <fulton@redhat.com>