Move test-operator CR name vars from vars to defaults
Currently, the names of the resources created by the test-operator role were defined
by variables in the vars folder of test-operator role. We want to move those vars
to defaults folder to enable their easier customization. The main reason why this name
customization might be useful is to be able to run tempest multiple times without doing
cleanup for every tempest run.