Enable running the test-operator role multiple times
It was not possible to run the test-operator role multiple times
without cleaning up all resources created by the first run, such
as the logs pod created for the logs collection.

This patch makes it possible to run the test operator role
multiple times without cleaning up the resources (e.g., after the
update in an update job). It is possible because:

- test-operator-logs pod now has unique name for each job run
- The variable cifmw_test_operator_*_name is now exposed. By
  specifying a unique name, we can guarantee no name collisions
  between the test-operator related resources (logs pods, instances
  of CRs)

If you want to run test-operator role twice within a single job
then make sure that for the second run you specify a value for the
cifmw_test_operator_*_name. For example:

  cifmw_test_operator_tempest_name: "post-update-tempest-tests"