Only consider ocps networks for network-values template
The common template for the network-values in ci_gen_kustomize_values
iterates over all the hosts in the cifmw_networking_env_definition and
stores in a map the relationship between a network and the interface.

For this to work, the ocp nodes must be the last ones, since if there
are other nodes after them in the cifmw_networking_env_definition,
for example nodes named osp-*, and they use a different network
interface, the resulting nncp will fail when applied.