Do not run control plane testing while openstackclient is updated.
To test the control plane during update we leverage the
openstackclient pod.

While the control plane is updated, the openstackclient pod is updated
as well. That means that during a short period of time (less than 2
seconds in the tests we have run) the openstackclient might be
unavailable.

To prevent false positive we:
1. make sure the error was not linked to the openstackclient shutting
down (code 137)
2. or, verify if the error message match the expected one for such
error.

Closes: https://issues.redhat.com/browse/OSPRH-10546