DNM Check binding dnsmasq listen-address to 0.0.0.0
There is an issue, that the DNS requests are redirected to the cloud
provider to resolve domain: api-int.crc.testing. Normally it is right
that it want to resolve, but the dnsmasq should take the query and
reply, however this is not the case.
Let's set the dnsmasq listen-address to be accessible to all addresses
on that host.