Don't rely on all-group.yaml for adoption osp role
Do not rely on reading the all-group.yaml file to parse all the vms
groups. Usually this would be correct, but since  we deploy ocp
separately from the osp vms, if the ocp playbooks runs afterwards it
rewrites the all-group file, leaving out the osp groups. This is becuase
in [1] we have to make the ocp deployment unaware of the rest of vms, to
avoid issues with config-drive.

[1] https://github.com/openstack-k8s-operators/ci-framework/pull/2445