[Shiftstack]Support proxy
We want to run the shiftstack role inside the cifmw.general collection
to run the shiftstack validation in an already deployed RHOSO.

For accessing it from the outside, we'll require to run the
kubernetes.core modules towards the underlying OCP through a proxy. That
is currently supported in that collection and we just need to set the
proxy parameter.

If the variable cifmw_shiftstack_proxy is not set, the variable is then
omitted, so backward compatibility is guaranteed.