Enable sushy-emulator in create-infra.yml
For adoption of the ironic service we need a Virtual BMC for spinning up
test workloads on the OSP 17.x cloud, and to validate workload
operations after adoption.

This change ensure the controller is started at the end of create-infra,
and does the neccecary steps to apply and validate the sushy-emulator
running as a POD on the controller node.

I have switched the use of `ansible.builtin.package` to
`ansible.builtin.dnf` in some places to avoid issues with hostvars
lookups raising undefined variable errors.