[test-operator][tobiko] Change downloaded images path
This patch changes the default path for downloaded images used for
tobiko testing from ~/.downloaded-images/ubuntu-minimal to
/usr/local/share.

This change is needed because we have to free /var/lib/tobiko from any
files that are present prior the container execution. This allows us to
mount ephemeral-storage to /var/lib/tobiko and use this directory as the
only point where write actions are allowed when readOnlyRootFilesystem:
true.

The location used for the downloaded images was changed with this [1]
commit in tcib.

[1] https://github.com/openstack-k8s-operators/tcib/pull/233