[kustomize_deploy] Make wait conditions optional
In some ocassions, like when applying more than one nodesets the user
may not want to wait for the first one to be applied and wait for all of
them at once when all are applied.
This change is about not forcing the user, for no reason, to run a
command to wait for a condition after some `oc apply` has been done.