Add the steps to reboot the computes after update.
This sequence implements reboot of the compute nodes after the update.

If one or more instances have been created they will be live-migrated
to others instance before the reboot and migrated back to the original
hypervisor after the reboot.

Some basic sanity checks are performed after the reboot and before the
migration back to ensure that the necessary services are up and
running.

Closes: https://issues.redhat.com/browse/OSPRH-8937