Simplify the role
Lets take all the dependencies on a running lunaclient VM
out.  We suspect that with the ip checking turned off, we can use
any subject in the client certs anyway, as long as the cert is uploaded
to the HSM (and registered to a client).

This simplifies things considerably - and makes the role more
reusable by a deployer as a standalone role.

Instead, we'll need the relevant bits uploaded somewhere for the
deployer to retrieve.