Add hypervisor param to shiftstack playbook execution
For being able to resolve shift-on-stack apps endpoints (routes)
hostnames from RHOSO pods it's required to add the DNS record entry
in then dnsmasq service running on the hypervisor.
It needs to be done after shift-on-stack cluster is installed as
the installer is creating the floating IP for the apps endpoints,
and the FIP is not known in advance.
For that purpose the cifmw_target_host param is passed so the playbook
running on the shiftstackclient pod is able to reach the hypervisor.