Log Ceph health and do not fail on HEALTH_WARN
The Ceph Upgrade tasks in the cifmw_cephadm role
will fail before the upgrade starts if the health
status is warn or error.

This patch changes it so that the upgrade only fails
if the cluster is in health error.

We have had the job fail in CI but we do not know why.
The task should log the Ceph health before starting the
upgrade so that CI results will give the job owner more
insight into why the job failed.

Signed-off-by: John Fulton <fulton@redhat.com>