compress text artifacts
This change proposes to compress text files where size above 1MB.

Toolings/libs/browsers are able to deflate on the fly, thus for
fair/substainable use of storage better to compress.

Some investigation shown that this simple step can save hundreds
of MB of data for a build. In the case of a CI artifacts storage
system and network bandwidth usage that non negligible.