Fix adoption_vars, only one defaultRoute, no ocbpm net
With PR# 2548 support was added for multiple edpm nodesets, these may
not have the same networks attached - i.e Compute and Networkere nodes
have different network needs. When introducing this support the
`networks` was made dynamic, and the `defaultRoute` was set for every
network.

Only one network can have the `defaultRoute`, this caused an error when
deploying the openstackdataplanenodeset.

Restore the behaviour prior to PR# 2548 - which would only set
`defaultRoute` for the `ctlplane` network.

Also, filter the cifmw_libvirt_manager_pub_net when generating networks
for edpm_nodes. cifmw_libvirt_manager_pub_net (typically named ocpbm) is
a ci-framework "management" network and should not be included in the
EDPM nodeset configs.

Jira: OSPRH-1301