[edpm_image_build] Copy certificates based on regexes
To avoid assuming the source certificate to copy into the
diskimage-builder has a specific name just use the system CAs chain and
allow the user to provide some regexes to select which CAs should be
copied.

The new module can read a PEM file (that can list many certs) and,
optionally, filter the certs by OU or CN regexes.