Create default RGW realm before rendering the spec
With Ceph Squid RGW fails to deploy because the default realm is
missing. In particular it fails with the following error:

'''
0 rgw main: failed to load realm: (2) No such file or directory
-1 Couldn't init storage provider (RADOS)
'''

that is clearly referred to a missing default realm.
This used to work well (probably because of a Ceph default) but
it started failing from Ceph Squid+ (RHCS 8).
This patch creates a default realm beforehand, and it is valid
both with RHCS 8 and previous releases.

Jira: https://issues.redhat.com/browse/OSPRH-10666

Signed-off-by: Francesco Pantano <fpantano@redhat.com>