[openshift_setup] Add support for overriding dependent operators CatalogSource
This is needed to easily test pre-release versions of OCP. OpenStack
depends on a few operators, these operators don't easily expose
pre-release versions for installation. A hack like this patch allows a
user to create a new CatalogSource for the current GA version of OCP and
patch the dependent operators to install from it.