adoption_osp_deploy : br-baremetal ovs_bridge
Change the NIC config template for OSP 17.1 deployment to use a
ovs_bridge for the br-baremetal interface.

Fixes issues in OVS wiring, when using linux bridge the patch to br-int
was not created and tap interfaces (qdhcp) was not correctly wired.

This also aligns the configuration used with the documentation.