unialpha adopt: provider-physical-network = ironic
To align with the physical network name used elsewhere:
* hooks/playbooks/ironic_network.yml
* reproducer/tasks/generate_bm_info.yml
* openstack-k8s-operators/architecture/examples/dt/uni01alpha/control-plane/service-values.yaml
* scenarios in gitlab ci-framework-jobs
  ...

Let's set the provider-physical-network to `ironic` in when creating
the provider network for the ironic provisioning/cleaning/rescue
network.

Also change it to use variables, like already done in:
  hooks/playbooks/ironic_network.yml

I will follow up with changes in data-plane-adoption to align there
as well.