[url_request] Use the system CA bundle before using the bundled one
requests by default uses the certifi CA bundle to validate SSL
certificates. If requests was installed using the system package manager
the package may point to the system CA bundle but if not it points to
certifi that may not have all the CAs installed in the system.

This approach tries to use the system CA bundle if possible.