Add Ansible variable cifmw_openshift_setup_apply_marketplace_fix
On some images (especially on legacy CRC images e.g. 2.30.0), there
is no need to run the openshift-marketplace workaround which is
removing all the pods from the namespace.
That step might be problematic on old deployments, due the
ImagePullPolicy is set to "Always", so after removing the
pods when the credentials are vanished, pods would be not recreated.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>