Change the method of installing test-operator
This patch introduces a new way of installing test-operator.
Before, we wanted test-operator role to install the test-operator,
but now test-operator will be installed the same way as other
operators - using openstack-operator.

Additionally, there is a need to ensure we can change which
version of test-operator is used. This will be necessary for
having a way to prevent big issues by changing the versions
temporarily.