[kustomize_deploy] Add name field in architecture steps
Till now, to add a kustomize patch users refered to the step they wanted
to patch by a numeric ID or an ID autogenerated based on the output
build path. To give users more control, this change allows users to add
a name field in each step in the automation field making patching a bit
easier.