Change default of expectedFailuresList parameter
The expectedFailuresList parameter is used to define which tests
are expected to fail so that they are not taken into consideration
when marking the job's status. It is useful for specific scenarios,
such as flaky tests, but we do not want every job to have this
parameter by default. It should be used manually to make its
intended use more explicit.