Add overrides for ci-framework and repo-setup
There is no 18.0-fr2 branch in ci-framework nor in repo-setup.
This causes issues when running content-provider or other jobs
against the 18.0-fr2, e.g. in telemetry-operator.
Hence, the override is proposed.

Additionally, after studying the Zuul documentation, we found
that for repositories with multiple branches there is implicit
branch matcher applied to each job and project-template definition,
hence we need to explicitly allow the jobs in other branches.