Use zuul clonned repos for upstream in build_openstack_packages role
In build_openstack_packages role for building rpms from gerrit and
github, we are clonning the repo and checking out specific change.

With these tasks, we are not able to build dlrn rpms if we have multiple
prs/crs from same project.

But we want to build rpms from all changes.
In order to fix, Zuul knows how to checkout proper repos
with changes under test and from Depends-on.

DLRN always looks for clonned repos in DLRN data directory.
In order to use zuul clonned sources with DLRN, We are creating
symlink to dlrn data project directory and let's dlrn do the job for
upstream repos only.

Note: we reverted the similar pr[1] here as original change was lacking
proper condition[2] leading to breaking downstream build openstack
packages role.

By adding proper conditional for upstream, it fixes the issue.

Links:
[1]. https://github.com/openstack-k8s-operators/ci-framework/pull/2818
[2]. https://github.com/openstack-k8s-operators/ci-framework/commit/f79ca6f89ae08948f74e34fdd32a691a7cef9f06#diff-fedeaff036de20345e170c4f65374926975f17139c058369f2e909565054e1adR118-L134

Signed-off-by: Chandan Kumar (raukadah) <raukadah@gmail.com>