DNM Moving cifmw-crc-podified-edpm-baremetal job to use just IBM cloud
From time to time we are out of the resources in one of our cloud
providers. The IBM private cloud provider is used rary, especially
for that job.
Let's try to use just this provider to unlock other CI jobs
that needs to be running on Vexxhost provider.

Also update the nodeset name to use CRC 2.48 - it was set
to wrong version by mistake.

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>