Monitor changes during update.
This collect changes in:
- Pod in openstack and openstack-operators namespace;
- containers on compute nodes;
- rpm on compute nodes;
- current step of the update;
and put it them in a timeline of event during the update process.

It creates a file `{{ cifmw_update_artifacts_basedir
}}/update_timeline.log` which is usually
`/home/zuul/ci-framework-data/tests/update/update_timeline.log` where
all the changes detected are logged in a `diff` format:

```diff
-keystone-cron-29050981-jvphr Succeeded registry.redhat.io/rhoso/openstack-keystone-rhel9a256:3fa3d690c31e9ec3146042609671f570cc6522a8ac68f87b5a82cc447cfd698d 2025-03-27T07:01:00
+keystone-cron-29051161-vt4zr Succeeded registry.redhat.io/rhoso/openstack-keystone-rhel9a256:3fa3d690c31e9ec3146042609671f570cc6522a8ac68f87b5a82cc447cfd698d 2025-03-27T10:01:00
```

Here we have change in the stage and a restart of the keystone-cron
container on the compute.

Furthermore during initialization of the logs, it collects the local
time on all host to make it easier to find the time delta between the
events on the controller-0 and the different nodes of the deployment.

The monitoring is done every
`cifmw_update_resources_monitoring_interval` seconds which is 10 by
default.

The molecule testing has been updated to adjust to the interface,
where monitoring initialization is done on its own and then the update
run.

Resolves-Part-Of:  [OSPRH-15060](https://issues.redhat.com//browse/OSPRH-15060)