New OLM parameter for managing deployment of installPlan approval.
A new parameter, `cifmw_ci_gen_kustomize_values_installplan_approval`,
is introduced for overriding OLM installPlanApproval values in
kustomize. It can be set to `Automatic` or `Manual`.

When set to `Manual` it has to be used in conjonction with
`cifmw_ci_gen_kustomize_values_deployment_version` as it won't work if
the the later is not set.

It shouldn't be necessary to set this to `Automatic` as it's the
default of operator installation.

Depends-On: https://github.com/openstack-k8s-operators/ci-framework/pull/2881

Resolves: [OSPRH-15054](https://issues.redhat.com//browse/OSPRH-15054)