Update CRDs when changing test-operator version
This patch updated the CRDs, if the user wants to patch the test
operator version. Until now the CRDs stayed unchanged even after
version change, so testing new test-operator parameters was not
possible. It is important to point out, that the update will not
be triggered in 90% of test-operator use cases.
It is primarily used in two situations:

1. When running checks for new PRs in the test-operator repository.

2. When updating the test-operator version in jobs to prevent
failures.

The approach isn’t the cleanest, but alternative solutions are
significantly more complex than the benefits they offer.