Removed variant of cifmw-molecule-tofu job
Multiple job definitions with the same name are called variants. These
may have different selection criteria which indicate to Zuul that, for
instance, the job should behave differently on a different git branch.
In this case, we have the same jobs, doing almost the same thing.

Removed duplicated job definition.