Add iscsi initiator file to the existing machineConfig template
We recently had an issue related to CIX-745 which sees glance failing to
upload an image when cinder is the backend. Apparently, it seems that
the current failure is related to the missing iscsi initiator in the
nodes where cinder-volume Pod is hosted. This patch updates the
machineConfig template to workaround the issue until [1] is merged and
promoted.

[1] https://review.opendev.org/c/openstack/cinder/+/946103

Signed-off-by: Francesco Pantano <fpantano@redhat.com>