Set concurrency for for tempest config without reference
There was many tests done to set the integer value for concurrency
parameter in tempest config. Even forcing the value to integer
does not meet with conversion in the end.
We thought that it is k8s module issue, but it seems not. It migth be
related to jinja2 according to issue [1].
Until the value is not parsed properly, we need to set it without
reference.

[1] https://github.com/ansible/ansible/issues/18095

Signed-off-by: Daniel Pawlik <dpawlik@redhat.com>