Optionally configure glance with cinder_volume_type=multiattach
With this patch, the hook that creates a multiattach volume type can
also configure that volume type as the `cinder_volume_type`, if
parameter `configure_cinder_volume_type` is set to true.

This is needed for octavia, because the amphora image that it creates
needs to be based on a multiattach volume. Otherwise, it fails to create
amphora VMs on different computes simultaneously, which is something
that happens when octavia is configured with ACTIVE_STANDBY.

OSPRH-16089
OSPCIX-768