dnsmasq: Fix start_v[4|6] is None
The task[1] has started to fail recently
on following error: "object of type
'NoneType' has no len(). object of type
'NoneType' has no len()"

I can see that cifmw_dnsmasq_network_definition in my case is:
    name: osp_trunk
    original_name: cifmw-osp_trunk
    ranges:
    - label: osp_trunk
      options:
      - option:dns-server,192.168.122.1
      - option:router
      prefix_length_v4: 24
      prefix_length_v6: null
      start_v4: 192.168.122.2
      start_v6: null
Which means that start_v6 is None and
that's defined but of NoneType so length
filter can't be used on it. The filter
does not need to be used tho It's enough
to condition, i.e. "if variable" (instead of
"If variable | length > 0") which
would be false if values is None, false,
0, empty string, dictionary or list.

[1] https://github.com/openstack-k8s-operators/ci-framework/blob/main/roles/dnsmasq/tasks/manage_network.yml#L35