Multinode EDPM CI job failing for nmstate provider
This change is to update the incorrect network config
template where used nic1 for dhcp and also used
br-ex controlplane config with default route.
so two default routes creating for nmstate provider
scenario and this is not happened for ifcfg
provider and only one default route created for nic1.
Proposed this PR is to fix incorrect network config.